App: Added type for the app state
This is only a start, in future, we will store the games properly etc.
(after we learn how to parse PGN and work out what we need to store at
each point). For now, we just have some placeholder state to prove that
we can store different state in different modes of the app.