App: Added the `ArbiterSetup` page
This page allows arbiters to enter the URL of a live game PGN link.

This is very incomplete, it is nothing more than a proof of concept that
we can access and modify the app state. In future, we will have to
implement the UI design as per figma, and track the state of the text
input