Chess Engine: Update `startGame` API
Previously, we were returning the starting board position as well as the
starting fen in a tuple - let's just return the starting board position.

This patch also changes the initial state for when we first enter
`GraphicalRecording` - previously, we would have a move history with a
whitePly property with only `startingFen` set.

It doesn't really make sense to have a move history before anybody has
made any moves - so we've ditched that, and now we start with an empty
history