Chess Engine: updated behaviour of `move` API
This patch also updates the `ChessPly` type.

A chess ply represents a move by a single player, so it doesn't make
sense to have this be optional - if a chess ply exists, it should have
the move that the player made

Having the outcome of a ply is much more informative than having the
starting fen of a ply.

The logic for the `move` function actually pushed the resultant fen as
the starting fen of the next move, that hasn't happened yet (which is
presumably the reason for squares being optional), so we may as well
ditch this and just store the resultant fen

The `move` function no longer pushes an extraneous move into the
history, and no longer depends on there being a dummy value in the
history at the start of the game